chore(index): move ScalarIndexPlugin trait into lance-index-core
ScalarIndexPlugin and all types it depends on now live in
lance-index-core, making the plugin interface available without
pulling in the full lance-index crate.

New modules in lance-index-core:
- scalar/registry: ScalarIndexPlugin, BasicTrainer, TrainingRequest,
  ScalarIndexLoad, ScalarIndexCacheKey, single_flight_open
- scalar/expression: ScalarQueryParser, IndexedExpression,
  ScalarIndexExpr, ScalarIndexSearch, ScalarIndexLoader
- scalar/seed: IndexSeedWriter, FragmentSeed, SEED_META_KEY_PREFIX
- progress: IndexBuildProgress and noop helpers
- registry: IndexRegistry trait (minimal plugin-lookup interface)

attach_registry now takes Arc<dyn IndexRegistry> instead of
Arc<IndexPluginRegistry> to avoid a circular dependency between
lance-index-core and lance-index.

ScalarIndexExpr::optimize() stays in lance-index as a
ScalarIndexExprOptimize extension trait because it references
SargableQuery, which has deep FTS dependencies that belong in
lance-index rather than the core crate.

All moved types are re-exported from their original paths in
lance-index for backwards compatibility.
