+++ rust/lance-index-core/Cargo.toml
@@ -12,6 +12,7 @@ categories.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+async-recursion.workspace = true
 async-trait.workspace = true
 arrow-array.workspace = true
 arrow-schema.workspace = true
@@ -28,6 +29,7 @@ prost-types.workspace = true
 roaring.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+tracing.workspace = true
 
 [lints]
 workspace = true
