+++ rust/lance-index/src/scalar/json.rs
@@ -36,7 +36,6 @@ use lance_core::{Error, ROW_ID, Result, cache::LanceCache, error::LanceOptionExt
 use crate::{
     Index, IndexType,
     metrics::MetricsCollector,
-    registry::IndexPluginRegistry,
     scalar::{
         AnyQuery, CreatedIndex, IndexStore, RowIdRemapper, ScalarIndex, SearchResult,
         UpdateCriteria,
@@ -47,6 +46,7 @@ use crate::{
         },
     },
 };
+use lance_index_core::registry::IndexRegistry;
 
 const JSON_INDEX_VERSION: u32 = 0;
 
@@ -397,7 +397,7 @@ impl TrainingRequest for JsonTrainingRequest {
 /// Plugin implementation for a [`JsonIndex`]
 #[derive(Default)]
 pub struct JsonIndexPlugin {
-    registry: Mutex<Option<Arc<IndexPluginRegistry>>>,
+    registry: Mutex<Option<Arc<dyn IndexRegistry>>>,
 }
 
 impl std::fmt::Debug for JsonIndexPlugin {
@@ -407,7 +407,7 @@ impl std::fmt::Debug for JsonIndexPlugin {
 }
 
 impl JsonIndexPlugin {
-    fn registry(&self) -> Result<Arc<IndexPluginRegistry>> {
+    fn registry(&self) -> Result<Arc<dyn IndexRegistry>> {
         Ok(self.registry.lock().unwrap().as_ref().expect_ok()?.clone())
     }
 
@@ -839,7 +839,7 @@ impl ScalarIndexPlugin for JsonIndexPlugin {
         true
     }
 
-    fn attach_registry(&self, registry: Arc<IndexPluginRegistry>) {
+    fn attach_registry(&self, registry: Arc<dyn IndexRegistry>) {
         let mut reg_ref = self.registry.lock().unwrap();
         *reg_ref = Some(registry);
     }
@@ -900,6 +900,7 @@ impl ScalarIndexPlugin for JsonIndexPlugin {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::registry::IndexPluginRegistry;
     use crate::scalar::{SargableQuery, TextQuery};
     use arrow_array::{ArrayRef, RecordBatch};
     use arrow_schema::{DataType, Field, Schema};
